requirements.txt
pyTelegramBotAPI
duckduckgo-search
python-dotenv
SpeechRecognition
pydub
Pillow
pytesseract